requirements.txt
streamlit
langchain
langchain-groq
langchain-community
pypdf
chromadb
sentence-transformers
python-dotenv